refactor(ui): split BackupManager into list/actions/schedule sub-components
